Fixed telegram commit logs
I just replaced the universal version with v7a version. Universal had the commit logs but it was too big for telegram so it didn't send. I didn't remove it because if you somehow managed to reduce the app size then it would work again. I also moved the "Upload last sha" to a higher position before compiling so it doesn't repeat the same commits if you push subsequently.